Add `unbork`
This is almost something broken with my PS1, maybe I should just run
`unbork` as part of that?

Either way, I sometimes break my shell when doing some dumb stuff that
might have printed some control characters. `unbork` seems to mostly fix
that without having to open a new shell, which is nice even if it's
definitely a bit of a hack.